# docs/swagger.yaml
basePath: /
definitions:
...
  models.City:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.SalesAssociate:
    properties:
      address:
        type: string
      area:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      discount:
        type: number
      email:
        type: string
      end_join_date:
        type: string
      id:
        type: string
      jenis_pembayaran:
        type: string
      join_date:
        type: string
      name:
        type: string
      phone1:
        type: string
      phone2:
        type: string
      updated_at:
        type: string
      website:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pustaka.com
    name: API Support
  description: REST API for Pustaka Digital Library Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pustaka Digital Backend API
  version: "1.0"
paths:
...
  /api/sales-associates:
    get:
      consumes:
      - application/json
      description: Retrieve all sales associates with their related city information
      produces:
      - application/json
      responses:
        "200":
          description: List of all sales associates
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all sales associates
      tags:
      - SalesAssociates
    post:
      consumes:
      - application/json
      description: Create a new sales associate entry
      parameters:
      - description: SalesAssociate details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SalesAssociate'
      produces:
      - application/json
      responses:
        "201":
          description: Created sales associate
          schema:
            $ref: '#/definitions/models.SalesAssociate'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new sales associate
      tags:
      - SalesAssociates
  /api/sales-associates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sales associate by ID
      parameters:
      - description: SalesAssociate ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SalesAssociate deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: SalesAssociate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a sales associate
      tags:
      - SalesAssociates
    get:
      consumes:
      - application/json
      description: Retrieve a single sales associate by its ID with related city information
      parameters:
      - description: SalesAssociate ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SalesAssociate details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: SalesAssociate not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a sales associate by ID
      tags:
      - SalesAssociates
    put:
      consumes:
      - application/json
      description: Update an existing sales associate by ID
      parameters:
      - description: SalesAssociate ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated sales associate details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SalesAssociate'
      produces:
      - application/json
      responses:
        "200":
          description: Updated sales associate
          schema:
            $ref: '#/definitions/models.SalesAssociate'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: SalesAssociate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a sales associate
      tags:
      - SalesAssociates
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"